fixup! refactor(calendar): update to new conventions
