+++ packages/components/src/Calendar/LegacyCalendarRange/LegacyCalendarRange.tsx
@@ -6,7 +6,7 @@ import {
   Matcher,
 } from "react-day-picker"
 import { ArrowBackwardIcon, ArrowForwardIcon } from "~components/Icon"
-import { legacyCalendarRangeClasses } from "../CalendarClassNames"
+import { legacyCalendarRangeClasses } from "../calendarClassNames"
 import { DayOfWeek } from "../enums"
 import { isInvalidDate } from "../utils/isInvalidDate"
 import { isValidWeekStartsOn } from "../utils/isValidWeekStartsOn"
